requirements.txt
flet
mysql-connector-python
bcrypt
python-dotenv